Open heart surgery on Rtl for x64.
It's mostly refactoring, it just looks really messy. Here's the problem:
on x64, the SSE floating point registers are architectural, and part of
the calling convention for passing/returning doubles. In the kernel, we
can't have the compiler using floating point registers whenever, since we
have to explicitly manage the context in those registers. So we need to
compile kernel code with -mno-sse. So the problem is when the printf
formatter does va_arg(va_list, double), that double is pulled from a
floating point register (due to the calling convention), but the compiler
correctly complains that it was being compiled with -mno-sse. The same
printf formatter is used in both kernel mode and user mode, which presents
a problem.

So the big thing this change does is pull out the double printing and
va_arg(double) into a separate file. Well, two really: one for kernel mode
that does nothing, and one for user mode that does something. This is good
because it shears off a bunch of useless float print formatting code from
the kernel, and solves an earlier problem I was having with -flto where
the float code got inlined up too far and was touching floating point
registers even when no float specifiers were in the format string.

We already have separate user and kernel pieces for Rtl, so that was a
natural place to put the float printing code. It means that krtl and urtl
are now a more internal part of Rtl, so to include rtlp.h I needed to move
urtl from osbase into lib/rtl. This is a more logical place for it anyway,
and it doesn't link against libminocaos since it's a static library, so
the dependencies work out.

It also means that the rtlc library needs this new pdouble.c file, which
meant moving rtlc into the new lib/rtl/urtl.

In pulling out the print double code, it also occurred to me that there
didn't really need to be a wide version if the printf emitter was a
function that could take either a wide or regular character. So I was able
to recombine a lot of print.c and wprint.c, including the integer and
float printing code. This would get more complicated if we wanted to allow
printf to emit many characters at once, but it's not insurmountable.

After all this, Rtl still doesn't really compile on x64, since I don't yet
have a way of adding -mno-sse CFLAGS to "sources in kernel mode". I'll need
to add some sort of identifier to the build files so they can identify one
way or the other, but that's a sprawling change so I want to do it
separately. Plus I have other changes backed up here that I need to flush.