Add 32-bit library build to x64.
The transition to long mode on PC/AT systems will happen while
transitioning between the boot manager and OS loader. This makes it easy
fo the boot manager to load both 32 and 64 bit OS loaders. It draws a
clear line where the MBR, VBR, Fatboot, and boot manager are all 32-bit,
and then the OS loader forward are 64-bit. The boot manager is also a
better place for the complexity of setting up page tables and switching to
long mode than Fatboot since it can be debugged, and is not file-system
specific.

The challenge is that the boot manager links a bunch of libraries in,
which all have to be compiled in 32-bit mode. This change adds those
32-bit library builds under x64.

This change may not work without the next couple of changes that come
after it, but it should be relatively close to isolated.