Refactor common definitions out of x86.h.
There was a lot of duplication between x86.h, x86.inc, x64.h, and x64.inc.
This change refactors all the preprocessor definitions common to all 4 of
these files out into a new x86defs.h, which can be included from both
assembly and C.

I originally also had an x86comm.h, which was a file containing C
(non-assembly) definitions common to both x86.h and x64.h, but there
turned out to be none of those. If a bunch of those crop up consider
adding that back in.