Kernel changes to get x64 boot manager to build.
This change contains a couple of bug fixes and build fixes needed to build
the 32-bit boot manager on x64. It also adds the x64 side of KD, though
this is currently untested. This change is basically "changes that are not
the boot apps themselves needed to get the boot manager compiling".