Boot app changes to compile x64 32-bit boot manager.
This commit contains changes to the boot manager and OS loader needed to
transition between to long mode. So far only PC/AT compiles, and because
of the -mno-sse issue I haven't been able to run it yet, so the trampoline
code is untested.

Most of this change has to do with de-pointerizing the
BOOT_INITIALIZATION_BLOCK structure passed between the boot manager and OS
loader. This can't have pointers because the two apps will disagree on
what size a pointer is. Now all the pointer fields are 64-bit integer
sizes.