Setup/build changes to create a skeletal x64 image.
This change adds a temporary setup platform called pc64 that contains only
the couple of files that build right now for x64. The goal is to get into
the 32-bit boot manager and then into the long mode OS loader. Also added
is some build infrastructure in the images directory.

For the makefile based build, you have to know what directories to make
and make them manually. For the mingen/build.ck based build, the image is
created normally by just running 'ninja'. The new build system really is
better, we should switch to it soon if I can get up the nerve to abandon
make.