Refactor build to identify kernel files.
x64 files compiled into a kernel, firmware, or boot environment need to
be compiled with -mno-sse and -mno-red-zone. User mode libraries and files
do not need this.

This change refactors the build so that binaries and libraries that build
for the kernel are identified as such.

With this change I can now jump to long mode with a test function that
simply returns and backs out to protected mode. First 64-bit code running!