Got into the OS loader debugger on x64.
With this change I'm now in the OS loader debugger running in x64 long
mode. In fact, I can even bounce down to 16-bit real mode and make
firmware calls.

This change spread out a bit because I did some refactoring of GDT
definitions. I think in x64 the GDT can just be defined as a static table
in C, and doesn't need functions to set up and manage descriptors. I had
to move some definitions around in order to make that easier to
accomplish.