Add vmware mouse support.
VMWare doesn't seem to enumerate a PS/2 mouse, or at least it doesn't
create a PNP0F*3 device. But they can create an EHCI controller with a
generic HID device behind it that happens to be a mouse. With some fixes
and recognition of that device specifically (since it does not enumerate
itself as a boot mouse), the mouse works in VMWare.