Add OSS device routing.
This changes gives the user the ability to select which external
input/output device to use. A sound core device may be wired to multiple
external devices (e.g. a speaker and a headphone jack). OSS provides a set
of ioctls to list which external devices (or routes) are available by name
and to select which named device to use by its index in the list. OSS does
not standardize on the names of these routes, but Minoca OS will.  For
example, headphone jacks will always be named "Headphone".