Fix e100 configuration space initialization.
A bug was introduced to the e100 driver when support for promiscuous mode
was added. Promiscuous mode is enabled by modifying a bit in the
configuration space with the e100 configure command. As there is no way to
perform a read-modify-write operation on the configuration space, enabling
promiscuous mode requires determining the correct values for the entire
configuration space. As part of this change, the act of initializing the
e100 device on initial driver load also started programming the
configuration space from scratch. Those initial values were not quite
right and left all e100 devices in half duplex mode. The correct
configuration enables the device to examine its full duplex pin to
determine if full duplex mode is supported. In half duplex mode, a 100Mbps
e100 runs closer to 10Mbps.

This change also adds PHY device detection to determine if an i82557 chip
is using a MII or the i82503 PHY.