Compiling x64 kernel.
I haven't run any of this code since I haven't made it through the loader
yet, but this change fills out many of the missing kernel functions for
x64 with a first stab. The kernel now links for x64, which should allow me
to make more progress through the loader.