Remove artifacts after native test.
Native test was leaving around the minoca-bin-xxx.tar.gz, and the packages
directory. Though this didn't accumulate since the next build would
overwrite the same place, it was causing problems on ARM machines, as
apparently we're close enough to maxing out our disk space to need this
space back.

This change removes those artifacts that would be uploaded in Native
Pilot, but are just abandoned in Native Test.