Use the distribution image for automation.
Before this change, we would build OS once to get distribution images,
then modify it for automation, then build it again to get a set of
automation images. This is time consuming, as some machines take a long
time to generate images, and with the increase in image size to 1GB it's
starting to be too much data to fit on split 32GB cards.

This change only builds OS once, and uses the distribution images to set
up the automation. It relies on the fact that msetup can install
additional files at install time. Now the install_build step will copy the
contents of /auto in during installation, as well as an init script.
That init script is responsible for installing requisite packages online,
and other automation setup steps (such as setting up SSH).