Fixes for getting into the x64 debugger cleanly.
This change removes the need for KdSkipMemoryValidation and
KdAvoidTimeCounter, and fixes a couple debugger issues.

The time counter was failing because rdtsc returns its value in eax:edx,
not rax. It needs a couple shift instructions to wrangle the value into
rax.

Memory validation was failing because the X64 page table macros were
slightly off.

The disassembler was missing the alternate group 15 encodings, which
include {rd,wr}{fs,gs}base.

The stack unwinder was failing because it turns out DWARF may report an
address size of 4 even for 64-bit architectures if it's successful in
squeezing all the addresses into 4 bytes. So we were doing a 4-byte memory
read instead of an 8-byte one when unwinding because we relied on the
DWARF address size to tell us the native word size.

Next up: something's wrong with HlpInterruptSetLineState.