Fix for strftime.
I had previously changed RtlFormatDate to return the number of characters
written NOT including the null terminator, but I failed to remove the
fixup strftime was doing. This was causing Python's strftime to return a
string without the last character, which caused client.py to fail to
update the current date, which caused the Raspberry Pi 2 to fail to build,
which is how I noticed there was an issue.