Fixes for taking address of packed struct member.
The new clang is persnickety about taking the address of a packed
structure member, as they realize that in some situations it could
result in a badly aligned pointer. For the most part they're being
overly cautious, but I'd like keep the warning enabled. This change
mostly uses locals or memcpy to avoid directly accessing potentially
unaligned pointers.