Got other processors started on x64.
We can now start other processors in Qemu on x64. Most of it involved
implementing ArSaveProcessorContext/ArRestoreProcessorContext and a couple
of fixes in ArInitializeProcessor.

This change also includes a couple ULONG truncations in the debugger. I
suspect there will be more of those.

There's also a fix to make the l modifier in RtlFormat (eg %lx) honor the
size of long (rather than LONG, which is always 32 bits).