Added PatchFile to Santa.
Santa now has a PatchFile and PatchSet class which are the complements to
the DiffFile and DiffSet classes. The new classes can apply patches to one
or multiple files based on a unified-diff style patch file.

Also noticed a couple of bugs in our diff implementation while testing it.
For text added to the end of a file with no deletions, our source line
number was off by one, and missing context.