Fix Chalk compiler bug.
The issue would show up with code like this:
var mylocal = 1;
var mydict = {"key": "value"};

1 ? 2 : 3;
mydict.key = mylocal;

The symptom is that mydict.key would end up with mydict in it, not 1. This
is because for "x.y" assignments we reload the local at StackSlots - 3. If
StackSlots isn't being tracked correctly, then the wrong local is loaded.
This happened in the case of the conditional expression, because the
compiler thought both 2 and 3 were pushed onto the stack, when in reality
only one expression or the other is evaluated.