Fix lzma buffer flow problem.
The Chalk lzma module was asserting that after encoding if the output is
not full then all input data should have been consumed. This ended up not
being true in a corner case. If the buffer is big enough, then the output
buffer is used directly. The encoder will internally stop when there's
less than 20 bytes of output space left. If there's still more input at
that time, then the encoder was quitting with both input data and a little
output space left.

The solution is to go back to our pre-allocated output buffer for the last
bit of input, then copy the odd bytes to the sliver of output remaining.
The tradeoff is that the remainder of the output from that input will now
have to be copied. But if the caller was trying to avoid that extra copy,
they should have supplied a bigger output buffer.