Fix for sh redirection failure quit.
The shell was considering failure to apply redirections to be a fatal
error, and exiting the shell entirely. Really it should have just
considered that command to be a failure, and moved on in the pipeline.
This change sets the last error code to 2, skips the command whose
redirections failed, and returns TRUE from the function.