Fix for Chalk JSON encoder corruption.
When encoding a JSON string, the encoder would sometimes reallocate its
buffer if there were escaped characters it needed to account for. It only
updated the buffer length at the very end. So the reallocated buffer did
not contain any of the partial progress the string encoder had made.

The solution is to update the length just before attempting to resize the
buffer.