x64: Fix "1 << x" truncations.
There are a bunch of places were we shift 1 by some amount and save it
into a UINTN. The problem is 1 defaults to an int type, so if you shift by
more than 32 you get weird undefined results. Add a L/UL/LL/ULL suffix
onto the end of "1"s that need it.

I originally noticed this because of an assert explostion in the MM block
allocator, and then went on a grepping mission for "1 <<". Note to self:
Be on the lookout for other constants getting shifted by an arbitrary
amount.