Only dispatch signals/timers on user mode faults.
Dispatching signals and checking for timers (insomuch as they may queue a
signal) require holding a process's queued lock. This can cause a dead
lock if the kernel faults while holding a process's lock. Skip checking
for timers and signals on kernel faults. If it's a user mode thread
faulting in the kernel, its signals and timers will be checked when it
exits back to user mode.