Santa: Add support for installing packages.
This change adds a few new classes:
Package - Manages the installation and status of an individual package. It
does not natively know where its control directory is.
PackageDatabase - Maintains a collection of PackageDepots. Supports
filtering for a selection of packages. It's a read-only interface to the
PackageDepots.
PackageManager - Maintains the list of packages installed on the system.
Coordinates and executes plans for installing and removing packages,
including resolving dependencies.

Also removed the Storage plugin infrastructure, since at this point it
would require too much back and forth between the modules, and the only
real choices were "directory" and "none". None got too tricky to implement
because it meant the extraction to storage was actually the real
installation, which meant duplicating a bunch of logic.