Fix fuzzy audio on Raspberry Pi.
While playing Mario on the RPI 3, sound seemed to be good 50% of the time
and when it was in the bad state, it never could catch up and correct
itself. Turns out that this was because the PWM audio driver was cheating
a bit on how it updated the sound buffer's position. It was assuming a 1:1
relationship between DMA interrupts and fragments. In truth, a single DMA
interrupt may account for multiple audio fragments if the interrupt
handler gets behind.

This change fixes the BCM2709 DMA driver to return the current DMA
transfer position when calling the DMA completion routine for continuous
transfers.