Chalk os: Add fork, waitpid, and get/setuid.
This change adds fork, waitpid, and a family of get/setuid related
functions for the Chalk OS module. This was done because in Santa needs to
do chroot-based containment during the build, but still needs to access
the outside world at the end of the build. On Windows, there's no
chroot-based containment, so fork failing is not a problem.