Minor chalk fixes.
This change fixes an issue with the cpio module where the permissions are
set to read-only before utimes is called, which at least on Windows causes
a permission denied error.

It also makes the "Builtin type does not implement get/set" messages a bit
more explicit with the actual type, which is helpful for debugging.