Add realms into Santa build, and install dependencies.
This change modifies the Santa package build process to run in a new realm
created specifically for the build. It also installs both build and host
dependency packages, storing host dependencies in a different directory if
cross compiling. This has only been tested on Windows, so I'll need to go
test on native Minoca.