Fix x64 debugger symbol truncation.
Frame unwinding had a bug where the CIE's address size parameter was being
used as the width of data to read from the target. There will probably be
more of these as I thought the address size corresponded to the target
machine's address size, not the smallest address size all the symbols
happen to fit into.