Kernel changes needed to get x64 booting to a shell.
This set of changes gets the x64 kernel all the way to the user mode
shell... usually. There appear to be some unknown mapping issues. On
single core I seem to lose a page table occasionally. On SMP I'm getting
triple faults. So not complete, but enough of a chunk of work that I'll
get those bug fixes in separately.

X64 doesn't do a "full" system call, and in fact I don't think any of the
architectures really need it. For the three system calls that use it, only
fork needs the full set of registers saved. So, I may change user mode
fork to save those registers themselves, then remove the system call
"full" type.