Add signal benchmarks to perftest.
This change adds three tests to perftest: one that sees how many ignored
signals can be raised in a given amount of time, one that sees how many
handled signals can be raised, and one that sees how many handled signals
with system call restarts can be raised.

These tests are good to get a sense of how fast we are at dealing with
signal primitives, although in the grand scheme of things signals aren't
usually one of the most performance sensitive primitives. It was useful to
me however while I was reorganizing the system call assembly and restart
mechanism.