Unified "full/sparse" system calls into one type.
This change unifies the "full" vs "fast" system call flavors into a single
flavor (equivalent to what used to be "fast"). Full system calls used to
be required by RestoreSignalContext and Exec so that they could restore
all trap frame registers. The full system call variant was also used by
Fork in order to be able to copy the full trap frame out to the child.

RestoreSignalContext and Exec are easy to solve by upgrading the trap
frame to "full" within the system call, and then checking in the system
call assembly and doing a slow full restore if the trap frame got upgraded
to full.

Fork was more complicated, since we needed to read, not write, the full
state. One option would be to check for fork explicitly in the assembly
and do a full save, but this seemed brittle if another system call comes
along with the same requirement. So instead we now ask user mode to
explictly save and restore the non-volatile registers across fork. This
means converting the OsForkProcess function into assembly so the
non-volatiles can be saved and reloaded across the system call.

On x86, I cheated and used the "full" system call mechanism anyway. Since
sysenter might not be supported, this mechanism is around anyway. The
assembly would have been ugly with __get_pc_thunk, _GOT_TABLE_OFFSET and
friends to get to the OsSystemCall global function pointer.

There's another reason to make the fork path assembly: vfork. When vfork
is implemented, the parent and child temporarily share the same stack.
Traditionally that means you cannot touch the stack in between making the
system call and returning all the way back to vfork's caller (because the
child will call other functions from vfork and trash that region). I've
added plumbing in the form of an extra fork parameter to be able to have
the kernel temporarily save and restore that particular region of user
stack. Getting this stack frame base difference requires assembly anyway.

I also rearranged the signal dispatch code and how it relates to system
call restarts. It turns out we don't need to pass around the system call
number and parameter to a bunch of functions. We can pull it out of the
trap frame.

Doing this all simplified the system call assembly significantly, and made
a lot more sense for x64, where the "full" concept was more awkward and
costly to pull off.