Return signal pending info from syscall C handler.
This change removes the code from the system call assembly that checked
the current thread signal pending state. This code was brittle as it
involved reaching through the PROCESSOR_BLOCK and KTHREAD structures, and
defining an enum (ThreadSignalPending). Now a boolean is returned from the
C system call handler, which already had the current thread, and that is
checked in the assembly instead. It also makes the logic into a consistent
(SignalPending != NoSignalPending), which was inconsistent in the
assembly (sometimes being SignalPending == ThreadSignalPending).