Got signals working on x64.
Only a couple of minor tweaks were needed to get signals working on x64.
I'm now able to run the perf test, though I did see resident set
accounting issues. I also noticed an issue with TLS relocations where the
sign was wrong. I copied it from the x86 relocation, so I'm not sure if
that one has the same problem or not.