Augment elfconv to handle 64-bit binaries.
Modify elfconv to be able to convert a 64-bit ELF binary into a 64-bit PE
file. Most of it was just changing 32s to 64s. Ideally this would be a
candidate for some sort of clever refactoring, however it's a pretty hokey
tool to begin with.

There's was a section present in the 32-bit file where if a relocation
section wasn't found, it looked for a dynamic section. I'm not sure if
that stuff is needed. When actually trying out a converted binary, nail
that down and either delete that section or make it specific to the 64-bit
architectures.