Fix IN6_IS_ADDR_V4MAPPED macro.
The first 5 words of an IPv4-Mapped IPv6 address are 0, followed by a word
of 0xFFFF and then the 32-bit IPv4 address. The old version of the macro
was checking to see if the first 7 words were 0 and the 8th was 0xFFFF.
This change fixes that a simplifies or cleans up a few other IPv6 macros.