Network address domains should match socket domain.
For functions like connect(), bind(), and sendto(), the supplied network
address's domain (a.k.a. family) should match the domain of the socket.
This was not being checked consistently for all routines and, if present,
was a hard-coded check for IPv4. This changes makes these checks more
generic and removes a few other IPv4 specific checks.

For the send() case, the checks are made in the protocol routines rather
than the netcore entry point or at the network layer. The protocol send
routines are in charge of copying the remote address out of the parameters
(typicall a copy from user mode), so the checks are performed as close to
the copy as possible.