Centralize network checksum routines.
In preparation for IPv6, which, unlike IPv4, requires UDP checksums, this
change creates a netcore API to checksum packet data. The algorithm is the
same for IP, TCP, and UDP, so it does not make sense to have three
implementations of the code.

As part of this, a pseudo-header checksum calculation routine was added to
the network interface. This allows the generic checksum code to not depend
on network specific definitions and implementations. TCP used to include
ip4.h in order to generate the pseudo-header checksum. Adding future
dependencies on IPv6 seemed wrong; the work is not farmed out to the
network layer.