Add initial IPv6 support (disabled).
This change adds the basic IPv6 network support. It is still disabled as it
depends on ICMPv6 (NDP) in order to function. That is the next step.

This version of IPv6 does not yet support fragmentation, multicast, or
extension headers. Those advanced features will be added after NDP is
implemented.