Convert link address entry list to an array.
This change converts the NET_LINK address entry list into an array of
lists, one for each socket network type. IPv4 address categorization needs
to get the local IP address in order to detect the subnet broadcast
address. With the forthcoming introduction of IPv6, each link would have
more than one link address entry, breaking the constant time local address
lookup for IPv4 address categorization.

As part of this change, DHCP is invoked for every link address entry in
the array of lists. This used to only happen for the first entry. That
said, DHCP isn't the correct protocol to invoke for all networks; the next
change will fix that.