Remove netcore's dependency on DHCP.
DHCP is IPv4 specific and netcore should not be calling it directly. This
change leaves link address configuration up to the network. In order to
make the network interface easy to find, each link address entry now stores
a pointer to it. A link address entry's network should be known upon
creation and never change. The network-specific address configuration
callback is then left to invoke DHCP, NDP, DHCPv6 or whatever protocol the
network depends on for address configuration.

Also, now that link address entry's store the network interface, finding a
link address entry for a local address no longer needs to take the network
as an argument. This simplifies a few function calls.