Move IPv4 and IPv6 sources into subfolders.
This change moves all the IPv4 and IPv6 specific protocols into their
respective subfolders. This will help organize the netcore folder better
as IPv6 is about to add files to handle ICMPv6, DHCPv6, NDP, and MLD. It
may also serve as a reminder to not have netcore directly depend on network
specific protocols (e.g. it used to directly call ARP and DHCP).