Fix for x86 sysenter regression.
The automation machines were hitting asserts and other issues because
interrupts were getting disabled in user mode. This was happening because
the sysenter code that was saving the complete trap frame was setting 0
into EFLAGS. Upon restoring after a signal was complete, this zero was
actually making it into eflags for user mode.