Fix PC single step issues.
Single stepping a process from user mode that had not yet set up its
signal handler was cauing kernel mode to take the single step exception.
This was not appropriate, as the single step was handled by the debugger
process.

Also fixed x64 double faults. The hardware pushes its own error code for
this exception, so there's no need to push a dummy one.

I was originally trying to debug an x64 issue with single stepping over a
syscall instruction in user mode. The FMASK MSR should have cleared TF,
preventing a single step exception in kernel mode. On Qemu this was not
happening, and ugliness was occurring because we were executing on the
user mode stack. It turns out this was a Qemu bug, so I'm not going to
worry about it.

I think there is still a problem with the system call refactoring on ARM,
but I cannot reproduce any problems with Qemu. I'll take a look at a real
machine when I get home.