Move multicast group accounting into netcore.
Before this change, IPv4 kept track of which multicast groups a socket
joined. It did this by allocating its own private IPv4 specific socket
structure. This information, however, will also be needed by IPv6 for
multicast support so it makes sense to centralize it into netcore.

This change also centralizes a link's accounting of which multicast
groups it belongs to. This is necessary as a complete list of IPv4 and
IPv6 multicast groups needs to be passed to the hardware in order to
update the multicast filters. It is not good enough to make add/remove
requests to the hardware due to how NICs implement multicast filters.