Fix for failing mmap(SHARED | ANONYMOUS).
Creating a shared anonymous section via mmap was failing because
MmSysMapOrUnmapMemory was handing IoSetFileInformation a kernel
FILE_PROPERTIES pointer but saying it was from user mode.

The fix is to say it's from kernel mode. This skips permission checks on
the truncate, but that's fine because we know the user has permission to
do that since they just created the handle.