Remove recently added fork resident set assert.
The assert was added while I was debugging x64 stuff. It was supposed to
say that every page except the shared user data page ought to be accounted
for by the image sections. This isn't quite true, since shared sections
get recreated rather than copied over. This means any pages in those
sections mapped in the source won't be mapped in the destination.

The weird thing is I was seeing this assert fire even when accounting for
those pages. It claimed destination was short by 1, but by the time I got
broken in to the debugger it's always correct. I suspect this has to do
with switching cores at the wrong time, causing the new value to be stuck
in another core's write buffer temporarily. But I can't explain that, since
it's manipulated with atomic adds, and context switching should also be a
full memory barrier. It leaves me with an uneasy feeling, but either way
the assert is incorrect.