Fixed ARM system call hang.
The recent system call refactoring caused a regression on ARM. The symptom
was that one core seemed to be going out to lunch in user mode. The issue
was in the code restoring pre-signal trap frames. It went to great lengths
to preserve SVC SP and LR from the trap frame created by that system call.
That used to be sufficient back when that was a "full" system call. Now
that it isn't, the SVC SP and LR are not saved by the system call
assembly, so uninitialized values were being restored for those values.

I believe this caused the thread to jump to another thread's stack, and
somehow get stuck spinning somewhere in kernel mode with interrupts
disabled.

This fix sets the address of the trap frame itself into SP. We assert that
this is in fact a pointer to the trap frame at the end of the kernel mode
stack, not some random stack allocated thing or trap frame elsewhere.

Note: in the days of using the full vs fast system calls, that bug still
existed, but was masked due to the fact that our libminocaos is well
behaved. Reducing the number degrees of freedom into the kernel
helped unmask that bug.

Chris also noticed R3 from kernel mode might get saved into the user mode
trap frame, leaking that register's contents in signal cases. Scrub R3 a
little earlier to fix that.