Fix a few IGMP bugs.
While attempting to reduce the amount of state stored by IGMP, an issue
with group reference counting was discovered. Basically, a join and leave
could have raced and resulted in the join touching a destroy group
structure. This change fixes that.

Unfortunately, the IGMP state cannot really be reduced, even though
netcore tracks multicast groups now. There are IGMP specific link and
group information and timers that netcore should not have to store.