Fix IGMP handling of unreported groups.
Certain multicast groups, like the all nodes multicast group should
never get reported when the host joins them or is queried by a router.
Handle this is a more generic way than specifically looking for and
assuming the existence of the all nodes multicast group when sending a
link-wide report.