Chalk rmdir on Windows: check directory contents.
Running rmdir in Chalk on Windows with a path that was not empty delayed
in that loop to handle cases where the directory was not quite cleared.
Check the directory explicitly before entering that loop to see if it's
empty, and return quickly if it's not.

This might backfire on me, since I'm not sure the case that loop was
protecting against won't show up in readdir.