Fix IPv6 header's version member.
The IPv6 header's version, traffic class, and flow label need to be
encapsulated by the same ULONG, but specificing bit fields creates the
wrong byte order. This needs to be done with a mask and byte swapping.

Also rename the IP6_ADDRESS structures Network member to Domain to fit
in with the other networks.