Finalize the x64 build images.
This change removes the crippled build support for x64. It now builds the
same files as x86, including the EFI boot manager.

I ended up needing to split out the pc, pcefi, and pc-tiny targets into 32
and 64-bit versions. The problem is that the "arch" variable in install.ck
is set from the platform recipe. So those platforms all set the arch as
"x86", which caused the EFI boot manager to get saved as BOOTIA32.EFI.
Now you can specify whether you want to set up the 32 or 64-bit build. On
a native setup target, it will guess the correct architecture based on the
pointer size of msetup itself. Given that we don't support 32-bit apps on
64-bit targets, this is sufficient.