Zero-extend small writes to ACPI buffer fields.
Say you (or in this case the Hyper-V ACPI firmware) try to write a
4-byte buffer into an 8-byte buffer field. We would fail to do that.
But the ACPI spec actually specifically says the smaller buffer
should be zero extended. Impressively thorough.

This change allocates a new temporary buffer of the needed size, and
creates a zero-extended copy. This should be rare enough to have a
negilible impact on perf.