Fixes to get x64 booting via UEFI.
Most of the work was in fixing up elfconv and the x64 linker script to get
things just lined up so. Overall this conversion thing is pretty fragile,
but given that there's not a lot more that really needs to go into the boot
manager it seems fine for now. Perhaps a better solution would be to see
if binutils could be configured with an additional target to create these
types of files directly.

There were also a couple changes to the assembly, where the @GOTPCREL
register needs to be dereferenced to get the value of the variable.

Finally, I had to add a prefix to the boot manager PCAT build, and stop
reusing bootman.o and bootim.o. This is because in PCAT builds, those are
compiled in 32-bit mode, whereas in the EFI boot manager those are 64-bit
object files.

I wasn't able to fully test the mingen build, because I hit the
strict-aliasing IPv6 compiler errors. I'll fix those sometime soon.

The system now boots, albeit a bit slowly due apparently to hammering the
disk, via x64 UEFI on VMWare.