sh fix for relative scripts on Windows.
Running a command like "autoconf" was failing on Windows even though
autoconf was in the path. The Win32 exec code was reading the shebang and
replacing the command appropriately, but left the old argv[0] relative. So
the command would end up being something like "sh autoconf" rather than "sh
/path/to/autoconf" like the shell found.

This change replaces what was argv[0] with the real path of the script,
so that whatever is interpreting it can find it.