Grep: use binary mode.
Grep on Windows was returning EOF early for binary files (like a .exe).
This caused libtool to end up installing the lt wrapper executables
instead of the final executables, because its grep of the exe for '%%%
MAGIC exe variable %%%' was failing.

Open all files as binary files, and then trim the line for the trailing \r
to keep things working as they were.