Santa: basic version resolution and base packages.
This change contains a bunch of bug fixes needed to get Santa building the
toolchain on Windows. It adds a set of base packages that are installed in
each build container.

It also adds some basic versioning resolution, so that a package can
depend on automake=1.12.6 or automake>=1.15. Doing actual version graph
resolution is NP-hard. This version will basically only install an older
version of a package if 1) required by a dependency specification, 2) not
already installed, and 3) a newer version is not already in the add plan.
This simplifies the problem significantly, at the cost of bailing out in
some cases where the newest package is not the best. Overall a bit cheezy,
but it should work for most cases.