Bootstrap build fixes.
The mingen build for libcrypt.so was missing -nostdlib, which is necessary
because libcrypt.so is built during the compiler bootstrapping process,
and crti.o and friends are not yet available.

Also fixed an issue where Chalk is not actually in the path during
bootstrapping. Call out its path explicitly.