Add support for NDP's Duplicate Address Detection.
This change adds duplicate address detection, which is all that is needed
to declare a link-local IPv6 address unique. DAD requires NDP's neighbor
solicitation and advertisment messages. It sends solicitations and if no
advertisements are received within the default timeout, an address is
unique. If an advertisement is received, then the address is not unique
and the system must not use it.