TLS: Fix initial copy null pointer dereference.
OsDynamicLoaderMain was initializing the main thread's TLS image for any
image in the initial load that had a TLS image size. The problem is we
only actually allocate space for those images that have the static flag
set. So for any image that uses the general dynamic model but happens to
be loaded in the initial load, OsDynamicLoaderMain tries to copy into a
NULL segment.